Add files via upload
This project is a simple calculator program written in C that performs basic arithmetic operations using a switch case. The calculator takes user input for two numbers and an arithmetic operator (+, -, *, /) and outputs the result. The program runs in a loop, allowing continuous calculations until the user chooses to exit by pressing x.